requirements.txt
anthropic>=0.40.0
streamlit>=1.38.0
pdfplumber>=0.11.0
python-docx>=1.1.0
